MyCell改为了CellBehavior 调整了SoldierPrefab
