--- HOMG_Unity/Assets/Scenes/SampleScene.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9fc0d4010bbf28b4594072e72b8655ab
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
