--- HOMG_Unity/Assets/Scenes/map 1.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: debc109b6da167f4caec79bf82638ff8
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
