--- HOMG_Unity/Assets/Scripts/Common/MyCell.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a1a4b71dcba63864fada3fd0e40c788e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
